Pin eslint version in CI (#93)
CI is currently red in this repo if you try to open a PR or commit to main because we implicitly download the latest version of eslint (major version 9) which has since been deprecated and no longer supports the eslint config file format we still use (bc this repo is old).

If we pin the version to the latest major version 8, it finally works again so we can land PRs again.